+++ Integrated/BBM/ADC/Servo.h
@@ -5,5 +5,5 @@ private:
 	unsigned int servo2_gpio;
 public:
 	void SetServo(unsigned int _servo1_gpio, unsigned int _servo2_gpio);
-	void MoveServo(int paddledegree1, int paddledegree2);
+	void MoveServo(double paddledegree1, double paddledegree2);
 };
